requirements.txt
backoff==1.10.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.0
requests==2.24.0
six==1.15.0
thrift==0.13.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
opentelemetry-proto

